+++ uefi-raw/src/protocol/pci/mod.rs
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub mod io;
 pub mod root_bridge;