+++ uefi-test-runner/src/proto/pci/mod.rs
@@ -1,7 +1,23 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub mod io;
 pub mod root_bridge;
+
+use uefi::Handle;
+use uefi::boot::{OpenProtocolAttributes, OpenProtocolParams, ScopedProtocol, image_handle};
+use uefi::proto::ProtocolPointer;
 
 pub fn test() {
     root_bridge::test();
+    io::test();
+}
+
+fn get_open_protocol<P: ProtocolPointer + ?Sized>(handle: Handle) -> ScopedProtocol<P> {
+    let open_opts = OpenProtocolParams {
+        handle,
+        agent: image_handle(),
+        controller: None,
+    };
+    let open_attrs = OpenProtocolAttributes::GetProtocol;
+    unsafe { uefi::boot::open_protocol(open_opts, open_attrs).unwrap() }
 }