+++ uefi-raw/CHANGELOG.md
@@ -7,12 +7,12 @@
   `EFI_PCI_ATTRIBUTE_MEMORY`, `EFI_PCI_ATTRIBUTE_BUS_MASTER`,
   `EFI_PCI_ATTRIBUTE_EMBEDDED_DEVICE`, and `EFI_PCI_ATTRIBUTE_EMBEDDED_ROM`
   to `PciRootBridgeIoProtocolAttributes`.
+- Added `PciIoProtocol` and related types.
 
 ## Changed
 - **Breaking**: Use `PxeBaseCodeBootType` (newtype-enum) instead of `u16` for
 `PxeBaseCodeSrvlist::server_type` and for the `server_type` parameter of
 `PxeBaseCodeSrvlist::new`.
-
 
 # uefi-raw - v0.16.0 (2026-08-25)